requirements.txt


## Related to the Environment
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
click==8.1.8
docstring_parser==0.16
exceptiongroup==1.2.2
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
python-dotenv==1.1.0
PyYAML==6.0.2
rich==14.0.0
shellingham==1.5.4
smmap==5.0.2
sniffio==1.3.1
tomli==2.2.1
typer==0.15.3
typing-inspection==0.4.0
typing_extensions==4.13.2
